+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # Changelog
 
+## [0.0.3] - 2020-02-08
+### Added
+- Added tools panel
+- Added save/open functionality
+### Changed
+- Changed plugin system
+
 ## [0.0.2] - 2020-02-02
 ### Added
 - Added orbit camera control
@@ -7,5 +14,6 @@
 ## [0.0.1] - 2020-02-01
 ### Initial release
 
+[0.0.3]: https://github.com/Perkovec/Vuxel/compare/v0.0.2...v0.0.3
 [0.0.2]: https://github.com/Perkovec/Vuxel/compare/v0.0.1...v0.0.2
 [0.0.1]: https://github.com/Perkovec/Vuxel/releases/tag/v0.0.1